1. Rename NavigationTypes.ts to navigationTypes.ts. 2. Create common/types folder and move navigationTypes.ts and userTypes.ts into it.
