+++ src/common/types/navigationTypes.ts
