+++ src/components/nearMe/UserListRenderer.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import { Box, ImageList, ImageListItem, Typography } from '@mui/material'
 import { LocationOn } from '@mui/icons-material'
-import { UserObjectType } from '../../common/userTypes'
+import { UserObjectType } from '../../common/types/userTypes'
 import Lightning from '../lightning/Lightning'
 
 type UserListRendererProps = {
